# Speed up _get_current_line() (take 2)